# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HTML to Image API",
    "description": "A REST API service that converts HTML content to images (PNG/JPEG) with support for mobile/desktop device simulation and custom sizing",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/convert/audio-to-video": {
      "post": {
        "tags": ["video"],
        "summary": "Convert audio to video",
        "description": "Converts audio file to video with waveform visualization, optional background image, and SRT subtitles",
        "operationId": "convertAudioToVideo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioToVideoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Video conversion completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioToVideoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Conversion failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/login/validate": {
      "post": {
        "tags": ["login"],
        "summary": "Validate platform login",
        "description": "Validates platform login status using cookies. If cookies are empty, automatically loads from local cookie files",
        "operationId": "validateLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginValidateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login validation successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginValidateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Login validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginValidateErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/login/platform/{platform}": {
      "get": {
        "tags": ["login"],
        "summary": "Get platform configuration",
        "description": "Returns configuration information for a specific platform",
        "operationId": "getPlatformConfig",
        "parameters": [
          {
            "name": "platform",
            "in": "path",
            "required": true,
            "description": "Platform identifier",
            "schema": {
              "type": "string",
              "enum": ["xiaohongshu", "douyin"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Platform configuration retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformConfigResponse"
                }
              }
            }
          },
          "404": {
            "description": "Platform not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          }
        },
        "required": ["error", "message"]
      },
      "Cookie": {
        "type": "object",
        "required": ["name", "value", "domain"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Cookie name"
          },
          "value": {
            "type": "string",
            "description": "Cookie value"
          },
          "domain": {
            "type": "string",
            "description": "Cookie domain",
            "example": ".xiaohongshu.com"
          },
          "path": {
            "type": "string",
            "default": "/",
            "description": "Cookie path"
          },
          "secure": {
            "type": "boolean",
            "default": false,
            "description": "Whether cookie is secure"
          },
          "httpOnly": {
            "type": "boolean",
            "default": false,
            "description": "Whether cookie is HTTP-only"
          },
          "session": {
            "type": "boolean",
            "default": false,
            "description": "Whether cookie is a session cookie"
          }
        }
      },
...
      "AudioToVideoRequest": {
        "type": "object",
        "required": ["audioUrl"],
        "properties": {
          "audioUrl": {
            "type": "string",
            "description": "Audio file URL"
          },
          "width": {
            "type": "integer",
            "default": 1080,
            "description": "Video width in pixels"
          },
          "height": {
            "type": "integer",
            "default": 1920,
            "description": "Video height in pixels"
          },
          "backgroundColor": {
            "type": "string",
            "default": "#000000",
            "description": "Background color (hex format)"
          },
          "backgroundImage": {
            "type": "string",
            "description": "Background image URL (optional)"
          },
          "waveColor": {
            "type": "string",
            "default": "#00ffff",
            "description": "Waveform color (hex format)"
          },
          "mode": {
            "type": "string",
            "enum": ["line", "point", "p2p", "cline", "bar"],
            "default": "bar",
            "description": "Waveform visualization mode"
          },
          "barWidth": {
            "type": "integer",
            "default": 4,
            "description": "Bar width for bar mode"
          },
          "barGap": {
            "type": "integer",
            "default": 2,
            "description": "Bar gap for bar mode"
          },
          "barSpacing": {
            "type": "integer",
            "default": 2,
            "description": "Bar spacing for bar mode"
          },
          "waveX": {
            "type": "integer",
            "description": "Waveform X position (default: 0)"
          },
          "waveY": {
            "type": "integer",
            "description": "Waveform Y position (default: 0)"
          },
          "waveWidth": {
            "type": "integer",
            "description": "Waveform width (default: video width)"
          },
          "waveHeight": {
            "type": "integer",
            "description": "Waveform height (default: video height)"
          },
          "srt": {
            "type": "string",
            "description": "SRT subtitle file URL (optional)"
          }
        }
      },
      "AudioToVideoResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "url": {
            "type": "string",
            "description": "Public URL to access the video"
          },
          "path": {
            "type": "string",
            "description": "Local file path"
          },
          "filename": {
            "type": "string",
            "description": "Output filename"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          },
          "waveX": {
            "type": "integer"
          },
          "waveY": {
            "type": "integer"
          },
          "waveWidth": {
            "type": "integer"
          },
          "waveHeight": {
            "type": "integer"
          },
          "barSpacing": {
            "type": "integer"
          },
          "srtUrl": {
            "type": "string",
            "description": "SRT subtitle file URL (if provided)"
          },
          "hasSubtitles": {
            "type": "boolean"
          },
          "copiedPath": {
            "type": "string",
            "description": "Path where video was copied (if applicable)"
          }
        }
      },
      "LoginValidateRequest": {
        "type": "object",
        "required": ["platform"],
        "properties": {
          "platform": {
            "type": "string",
            "enum": ["xiaohongshu", "douyin"],
            "description": "Platform identifier"
          },
          "cookies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cookie"
            },
            "description": "Cookies array (optional, if empty will load from local file)"
          },
          "options": {
            "type": "object",
            "properties": {
              "headless": {
                "type": "boolean",
                "default": true,
                "description": "Run browser in headless mode"
              },
              "timeout": {
                "type": "integer",
                "default": 30000,
                "description": "Request timeout in milliseconds"
              },
              "savePath": {
                "type": "string",
                "default": "./cookies",
                "description": "Path to save/load cookies"
              },
              "cookieFilePath": {
                "type": "string",
                "description": "Specific cookie file path (optional, if not provided will find latest)"
              }
            }
          },
          "cookieFilePath": {
            "type": "string",
            "description": "Specific cookie file path (alternative to options.cookieFilePath)"
          }
        }
      },
      "LoginValidateResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "platform": {
            "type": "string"
          },
          "platformName": {
            "type": "string",
            "description": "Display name of the platform"
          },
          "message": {
            "type": "string"
          },
          "cookies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cookie"
            }
          },
          "cookiesCount": {
            "type": "integer"
          },
          "savedPath": {
            "type": "string",
            "description": "Path where cookies were saved"
          }
        }
      },
      "LoginValidateErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "platform": {
            "type": "string"
          },
          "platformName": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "enum": ["validation_failed", "no_cookies_available"]
          }
        }
      },
...
      "PlatformConfigResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "platform": {
            "type": "object",
            "properties": {
              "key": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "displayName": {
                "type": "string"
              },
              "loginUrl": {
                "type": "string"
              },
              "checkUrl": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
...
}